+++ changelog.md
@@ -6,6 +6,7 @@
 > This version is not released yet and is under active development.
 
 - Add `extra-platforms` CLI command.
+- Add new `ALPINE`, `KALI`, `MANJARO`, `OPENWRT` platform definitions.
 
 ## [9.0.0 (2026-02-11)](https://github.com/kdeldycke/extra-platforms/compare/v8.0.0...v9.0.0)